move banner component to new lib layout
