+++ projects/storefrontlib/src/cms-components/content/banner/banner.component.html
