+++ projects/storefrontlib/src/cms-components/content/index.ts
@@ -1,0 +1,2 @@
+export * from './banner/banner.component';
+export * from './banner/banner.module';
