+++ projects/storefrontlib/src/cms-components/index.ts
@@ -1,3 +1,4 @@
 export * from './checkout/index';
+export * from './content/index';
 export * from './misc/index';
 export * from './user/index';
